feat(is-class): add isClass function to check if the input is a class
This commit adds a new file `is-class.ts` which contains the implementation of the `isClass` function. This function checks if the input is a class by using the `typeOf` function from the `typeof.ts` module. The `isClass` function returns `true` if the input is a class and `false` otherwise.

A new test file `is-class.test.ts` is also added to test the `isClass` function. The test file contains multiple test cases to check if different types of values are considered as classes or not. Each test case asserts the expected result using the `expect` function from the `vitest` library.

The purpose of this change is to provide a utility function that can be used to determine if a given value is a class or not. This can be useful in scenarios where it is necessary to differentiate between classes and other types of values.