feat(is-regexp): add `isRegExp` function to check if the input is a regular expression
This commit adds a new file `is-regexp.ts` which contains the implementation of the `isRegExp` function. This function checks if the input is a regular expression. It uses the `typeOf` function from the `typeof.ts` file to determine the type of the input and returns `true` if the type is `'regexp'`, otherwise it returns `false`.

A new test file `is-regexp.test.ts` is also added to test the `isRegExp` function. It includes multiple test cases to check if different types of inputs are correctly identified as regular expressions or not.